feat(ssl): Fake CA and key to generate certificates for development
I added documentation on how the CA can be used.
Also re-generated the org.crt and org.key for the clerk frontend.

When this CA is imported and trusted by the browser https://localhost/
Should have a "valid" certificate.
If this is not instantly the case after reset check again in a incognito
frame. Could be browser cache ;)